Split library
added initial certs
