+++ certs/cert/ec-cacert.pem
@@ -1,0 +1,13 @@
+-----BEGIN CERTIFICATE-----
+MIIB7TCCAZOgAwIBAgIUVIgTaTTwoQsb9wWSQ6Om5C8iQEYwCgYIKoZIzj0EAwIw
+TDELMAkGA1UEBhMCQVUxDDAKBgNVBAgMA1ZpYzEMMAoGA1UEBwwDTWVsMSEwHwYD
+VQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwHhcNMjIwNzE0MDcyMTEzWhcN
+MzIwNzExMDcyMTEzWjBMMQswCQYDVQQGEwJBVTEMMAoGA1UECAwDVmljMQwwCgYD
+VQQHDANNZWwxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDBZMBMG
+ByqGSM49AgEGCCqGSM49AwEHA0IABAEDnjo0jYHYWdQzK0ILV1TyVYw7PcEaaBNh
+xfIvRwJ9CskWOFjTjhqvrSr/u92yUjwwVDVPeDifikkKTQRjrwyjUzBRMB0GA1Ud
+DgQWBBR4JwqT+BmSgfSWtMXRrO9spFECtDAfBgNVHSMEGDAWgBR4JwqT+BmSgfSW
+tMXRrO9spFECtDAPBgNVHRMBAf8EBTADAQH/MAoGCCqGSM49BAMCA0gAMEUCIQDi
+GLelyElOFHmRSd6dq2K3bpPNdHnnHyL/rx640cpazwIgeJpSvdon1sAehGUt3DLQ
+G0RwhZuKDbCo8SWnqA+/vp4=
+-----END CERTIFICATE-----
