+++ certs/client.crt
@@ -1,0 +1,58 @@
+Certificate:
+    Data:
+        Version: 3 (0x2)
+        Serial Number: 2 (0x2)
+        Signature Algorithm: ecdsa-with-SHA256
+        Issuer: C=AU, ST=Vic, L=Mel, O=Internet Widgits Pty Ltd
+        Validity
+            Not Before: Jul 14 07:21:54 2022 GMT
+            Not After : Jul 14 07:21:54 2023 GMT
+        Subject: C=AU, ST=Vic, O=Internet Widgits Pty Ltd, CN=client01/emailAddress=first@first.com
+        Subject Public Key Info:
+            Public Key Algorithm: id-ecPublicKey
+                Public-Key: (256 bit)
+                pub:
+                    04:ba:c5:29:54:52:c5:e6:da:51:2f:24:b6:a4:25:
+                    16:2b:90:98:ca:5a:05:9e:77:5d:57:cf:31:05:c8:
+                    80:d5:95:f9:75:1e:c9:b0:e6:20:ac:67:fd:f7:2c:
+                    f4:96:42:cc:62:75:be:08:b0:2c:c2:a4:49:b4:6a:
+                    8f:5d:e7:cf:65
+                ASN1 OID: prime256v1
+                NIST CURVE: P-256
+        X509v3 extensions:
+            X509v3 Basic Constraints: 
+                CA:FALSE
+            Netscape Cert Type: 
+                SSL Client, S/MIME
+            Netscape Comment: 
+                OpenSSL Generated Client Certificate
+            X509v3 Subject Key Identifier: 
+                B9:0B:CC:D1:AE:28:51:1B:6C:A6:DD:BD:3F:74:47:B1:D3:B9:FC:04
+            X509v3 Authority Key Identifier: 
+                keyid:78:27:0A:93:F8:19:92:81:F4:96:B4:C5:D1:AC:EF:6C:A4:51:02:B4
+
+            X509v3 Key Usage: critical
+                Digital Signature, Non Repudiation, Key Encipherment
+            X509v3 Extended Key Usage: 
+                TLS Web Client Authentication, E-mail Protection
+    Signature Algorithm: ecdsa-with-SHA256
+         30:45:02:20:34:1a:27:68:c5:f4:f3:85:1f:b5:fd:0b:76:6f:
+         25:3a:49:26:0b:ed:04:62:7f:8d:20:27:9e:10:2b:60:00:f4:
+         02:21:00:e2:1b:a4:17:a1:fb:86:ef:6b:a5:23:64:9b:47:b6:
+         20:26:63:13:4b:60:7b:06:ff:4f:c9:89:c7:4d:f9:63:e2
+-----BEGIN CERTIFICATE-----
+MIICcjCCAhigAwIBAgIBAjAKBggqhkjOPQQDAjBMMQswCQYDVQQGEwJBVTEMMAoG
+A1UECAwDVmljMQwwCgYDVQQHDANNZWwxITAfBgNVBAoMGEludGVybmV0IFdpZGdp
+dHMgUHR5IEx0ZDAeFw0yMjA3MTQwNzIxNTRaFw0yMzA3MTQwNzIxNTRaMHExCzAJ
+BgNVBAYTAkFVMQwwCgYDVQQIDANWaWMxITAfBgNVBAoMGEludGVybmV0IFdpZGdp
+dHMgUHR5IEx0ZDERMA8GA1UEAwwIY2xpZW50MDExHjAcBgkqhkiG9w0BCQEWD2Zp
+cnN0QGZpcnN0LmNvbTBZMBMGByqGSM49AgEGCCqGSM49AwEHA0IABLrFKVRSxeba
+US8ktqQlFiuQmMpaBZ53XVfPMQXIgNWV+XUeybDmIKxn/fcs9JZCzGJ1vgiwLMKk
+SbRqj13nz2WjgcUwgcIwCQYDVR0TBAIwADARBglghkgBhvhCAQEEBAMCBaAwMwYJ
+YIZIAYb4QgENBCYWJE9wZW5TU0wgR2VuZXJhdGVkIENsaWVudCBDZXJ0aWZpY2F0
+ZTAdBgNVHQ4EFgQUuQvM0a4oURtspt29P3RHsdO5/AQwHwYDVR0jBBgwFoAUeCcK
+k/gZkoH0lrTF0azvbKRRArQwDgYDVR0PAQH/BAQDAgXgMB0GA1UdJQQWMBQGCCsG
+AQUFBwMCBggrBgEFBQcDBDAKBggqhkjOPQQDAgNIADBFAiA0GidoxfTzhR+1/Qt2
+byU6SSYL7QRif40gJ54QK2AA9AIhAOIbpBeh+4bva6UjZJtHtiAmYxNLYHsG/0/J
+icdN+WPi
+-----END CERTIFICATE-----
