+++ certs/client.csr
@@ -1,0 +1,9 @@
+-----BEGIN CERTIFICATE REQUEST-----
+MIIBOjCB4QIBADB/MQswCQYDVQQGEwJBVTEMMAoGA1UECAwDVmljMQwwCgYDVQQH
+DANNZWwxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDERMA8GA1UE
+AwwIY2xpZW50MDExHjAcBgkqhkiG9w0BCQEWD2ZpcnN0QGZpcnN0LmNvbTBZMBMG
+ByqGSM49AgEGCCqGSM49AwEHA0IABLrFKVRSxebaUS8ktqQlFiuQmMpaBZ53XVfP
+MQXIgNWV+XUeybDmIKxn/fcs9JZCzGJ1vgiwLMKkSbRqj13nz2WgADAKBggqhkjO
+PQQDAgNIADBFAiB8K545L6cq7O5/pFyZwCpgouXLpcsOE0rjrQBV74TYmQIhAMXO
+IY2cvARl0RUjVg/Gu/5qs/qcSKGXCfgDAgXZ4KQJ
+-----END CERTIFICATE REQUEST-----
