+++ certs/client.key
@@ -1,0 +1,8 @@
+-----BEGIN EC PARAMETERS-----
+BggqhkjOPQMBBw==
+-----END EC PARAMETERS-----
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIH/LRGScZ4opy4Jf3E5C4pg91BSLwpSwjlQzhgBnqu3uoAoGCCqGSM49
+AwEHoUQDQgAEusUpVFLF5tpRLyS2pCUWK5CYyloFnnddV88xBciA1ZX5dR7JsOYg
+rGf99yz0lkLMYnW+CLAswqRJtGqPXefPZQ==
+-----END EC PRIVATE KEY-----
