+++ certs/index.txt
@@ -1,0 +1,2 @@
+V	230714072127Z		01	unknown	/C=AU/ST=Vic/O=Internet Widgits Pty Ltd/CN=localhost
+V	230714072154Z		02	unknown	/C=AU/ST=Vic/O=Internet Widgits Pty Ltd/CN=client01/emailAddress=first@first.com
