+++ certs/index.txt.attr
@@ -1,0 +1,1 @@
+unique_subject = yes
