+++ certs/index.txt.attr.old
@@ -1,0 +1,1 @@
+unique_subject = yes
