+++ certs/private/ec-cakey.pem
@@ -1,0 +1,8 @@
+-----BEGIN EC PARAMETERS-----
+BggqhkjOPQMBBw==
+-----END EC PARAMETERS-----
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIDiPHv2VR2AzB/0ESffEHFJkEK5FFl8DUrmUT5U5aHbDoAoGCCqGSM49
+AwEHoUQDQgAEAQOeOjSNgdhZ1DMrQgtXVPJVjDs9wRpoE2HF8i9HAn0KyRY4WNOO
+Gq+tKv+73bJSPDBUNU94OJ+KSQpNBGOvDA==
+-----END EC PRIVATE KEY-----
