+++ certs/serial.old
@@ -1,0 +1,1 @@
+02
