+++ certs/server.csr
@@ -1,0 +1,8 @@
+-----BEGIN CERTIFICATE REQUEST-----
+MIIBGzCBwgIBADBgMQswCQYDVQQGEwJBVTEMMAoGA1UECAwDVmljMQwwCgYDVQQH
+DANNZWwxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDESMBAGA1UE
+AwwJbG9jYWxob3N0MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEv9Sf4/xA+4X/
+pU9qutMA6KLl67mNdn0X88v2BjUc4/tGq95gwgo80rYvJS7k+7UQHcyVE1mcYA78
+wfef04UTmqAAMAoGCCqGSM49BAMCA0gAMEUCIBMjCRnRMVFUbqXhV0FzP4ZLBgjE
+ALagtseZoGqLFp5jAiEA5+KM5JScj/SSwIqUrUFaCy2Wjc6rIW+PQJztuuHXReA=
+-----END CERTIFICATE REQUEST-----
