+++ certs/server.key
@@ -1,0 +1,8 @@
+-----BEGIN EC PARAMETERS-----
+BggqhkjOPQMBBw==
+-----END EC PARAMETERS-----
+-----BEGIN EC PRIVATE KEY-----
+MHcCAQEEIOxat4szLCndA43LyC1sXgSZ97p/BTdAU6oKzo5udyyYoAoGCCqGSM49
+AwEHoUQDQgAEv9Sf4/xA+4X/pU9qutMA6KLl67mNdn0X88v2BjUc4/tGq95gwgo8
+0rYvJS7k+7UQHcyVE1mcYA78wfef04UTmg==
+-----END EC PRIVATE KEY-----
