+++ certs/server.pem
@@ -1,0 +1,77 @@
+Certificate:
+    Data:
+        Version: 3 (0x2)
+        Serial Number: 1 (0x1)
+        Signature Algorithm: ecdsa-with-SHA256
+        Issuer: C=AU, ST=Vic, L=Mel, O=Internet Widgits Pty Ltd
+        Validity
+            Not Before: Jul 14 07:21:27 2022 GMT
+            Not After : Jul 14 07:21:27 2023 GMT
+        Subject: C=AU, ST=Vic, O=Internet Widgits Pty Ltd, CN=localhost
+        Subject Public Key Info:
+            Public Key Algorithm: id-ecPublicKey
+                Public-Key: (256 bit)
+                pub:
+                    04:bf:d4:9f:e3:fc:40:fb:85:ff:a5:4f:6a:ba:d3:
+                    00:e8:a2:e5:eb:b9:8d:76:7d:17:f3:cb:f6:06:35:
+                    1c:e3:fb:46:ab:de:60:c2:0a:3c:d2:b6:2f:25:2e:
+                    e4:fb:b5:10:1d:cc:95:13:59:9c:60:0e:fc:c1:f7:
+                    9f:d3:85:13:9a
+                ASN1 OID: prime256v1
+                NIST CURVE: P-256
+        X509v3 extensions:
+            X509v3 Basic Constraints: 
+                CA:FALSE
+            Netscape Cert Type: 
+                SSL Server
+            Netscape Comment: 
+                OpenSSL Generated Server Certificate
+            X509v3 Subject Key Identifier: 
+                C0:7B:89:83:F5:E5:B0:90:A1:BB:E2:7D:3F:01:DA:00:7E:17:0C:04
+            X509v3 Authority Key Identifier: 
+                keyid:78:27:0A:93:F8:19:92:81:F4:96:B4:C5:D1:AC:EF:6C:A4:51:02:B4
+                DirName:/C=AU/ST=Vic/L=Mel/O=Internet Widgits Pty Ltd
+                serial:54:88:13:69:34:F0:A1:0B:1B:F7:05:92:43:A3:A6:E4:2F:22:40:46
+
+            X509v3 Key Usage: critical
+                Digital Signature, Key Encipherment
+            X509v3 Extended Key Usage: 
+                TLS Web Server Authentication
+            X509v3 Subject Alternative Name: 
+                IP Address:127.0.0.1, DNS:localhost
+    Signature Algorithm: ecdsa-with-SHA256
+         30:45:02:20:63:62:92:5a:31:c6:44:78:48:d5:5a:8a:f7:19:
+         77:9d:4a:bf:98:53:d1:52:3e:ee:d4:c8:d8:76:66:1e:36:79:
+         02:21:00:98:c4:f2:03:c7:6f:e6:b9:72:25:c6:6a:67:52:b3:
+         65:14:dd:21:57:62:06:e0:09:c7:b0:f6:6d:db:3e:3b:61
+-----BEGIN CERTIFICATE-----
+MIIC0TCCAnegAwIBAgIBATAKBggqhkjOPQQDAjBMMQswCQYDVQQGEwJBVTEMMAoG
+A1UECAwDVmljMQwwCgYDVQQHDANNZWwxITAfBgNVBAoMGEludGVybmV0IFdpZGdp
+dHMgUHR5IEx0ZDAeFw0yMjA3MTQwNzIxMjdaFw0yMzA3MTQwNzIxMjdaMFIxCzAJ
+BgNVBAYTAkFVMQwwCgYDVQQIDANWaWMxITAfBgNVBAoMGEludGVybmV0IFdpZGdp
+dHMgUHR5IEx0ZDESMBAGA1UEAwwJbG9jYWxob3N0MFkwEwYHKoZIzj0CAQYIKoZI
+zj0DAQcDQgAEv9Sf4/xA+4X/pU9qutMA6KLl67mNdn0X88v2BjUc4/tGq95gwgo8
+0rYvJS7k+7UQHcyVE1mcYA78wfef04UTmqOCAUIwggE+MAkGA1UdEwQCMAAwEQYJ
+YIZIAYb4QgEBBAQDAgZAMDMGCWCGSAGG+EIBDQQmFiRPcGVuU1NMIEdlbmVyYXRl
+ZCBTZXJ2ZXIgQ2VydGlmaWNhdGUwHQYDVR0OBBYEFMB7iYP15bCQobvifT8B2gB+
+FwwEMIGIBgNVHSMEgYAwfoAUeCcKk/gZkoH0lrTF0azvbKRRArShUKROMEwxCzAJ
+BgNVBAYTAkFVMQwwCgYDVQQIDANWaWMxDDAKBgNVBAcMA01lbDEhMB8GA1UECgwY
+SW50ZXJuZXQgV2lkZ2l0cyBQdHkgTHRkghRUiBNpNPChCxv3BZJDo6bkLyJARjAO
+BgNVHQ8BAf8EBAMCBaAwEwYDVR0lBAwwCgYIKwYBBQUHAwEwGgYDVR0RBBMwEYcE
+fwAAAYIJbG9jYWxob3N0MAoGCCqGSM49BAMCA0gAMEUCIGNikloxxkR4SNVaivcZ
+d51Kv5hT0VI+7tTI2HZmHjZ5AiEAmMTyA8dv5rlyJcZqZ1KzZRTdIVdiBuAJx7D2
+bds+O2E=
+-----END CERTIFICATE-----
+-----BEGIN CERTIFICATE-----
+MIIB7TCCAZOgAwIBAgIUVIgTaTTwoQsb9wWSQ6Om5C8iQEYwCgYIKoZIzj0EAwIw
+TDELMAkGA1UEBhMCQVUxDDAKBgNVBAgMA1ZpYzEMMAoGA1UEBwwDTWVsMSEwHwYD
+VQQKDBhJbnRlcm5ldCBXaWRnaXRzIFB0eSBMdGQwHhcNMjIwNzE0MDcyMTEzWhcN
+MzIwNzExMDcyMTEzWjBMMQswCQYDVQQGEwJBVTEMMAoGA1UECAwDVmljMQwwCgYD
+VQQHDANNZWwxITAfBgNVBAoMGEludGVybmV0IFdpZGdpdHMgUHR5IEx0ZDBZMBMG
+ByqGSM49AgEGCCqGSM49AwEHA0IABAEDnjo0jYHYWdQzK0ILV1TyVYw7PcEaaBNh
+xfIvRwJ9CskWOFjTjhqvrSr/u92yUjwwVDVPeDifikkKTQRjrwyjUzBRMB0GA1Ud
+DgQWBBR4JwqT+BmSgfSWtMXRrO9spFECtDAfBgNVHSMEGDAWgBR4JwqT+BmSgfSW
+tMXRrO9spFECtDAPBgNVHRMBAf8EBTADAQH/MAoGCCqGSM49BAMCA0gAMEUCIQDi
+GLelyElOFHmRSd6dq2K3bpPNdHnnHyL/rx640cpazwIgeJpSvdon1sAehGUt3DLQ
+G0RwhZuKDbCo8SWnqA+/vp4=
+-----END CERTIFICATE-----
