+++ src/config.rs
@@ -1,4 +1,4 @@
-use crate::{ LoadBalancer,  Client, ServerGroup, HealthChecker };
+use crate::{ LoadBalancer,  client::Client, server::ServerGroup, server::HealthChecker };
 use std::sync::Arc;
 use rustls::server::AllowAnyAuthenticatedClient;
 use rustls::{self, RootCertStore};
@@ -17,29 +17,29 @@ fn load_certs(filename: &str) -> Result<Vec<rustls::Certificate>, Box<dyn std::e
         .collect())
 }
 
-fn load_private_key(filename: &str) -> rustls::PrivateKey {
+fn load_private_key(filename: &str) -> Result<rustls::PrivateKey, Box<dyn std::error::Error>>
+{
     let keyfile = std::fs::File::open(filename).expect("cannot open private key file");
     let mut reader = BufReader::new(keyfile);
 
-    loop {
-        match rustls_pemfile::read_one(&mut reader).expect("cannot parse private key .pem file") {
-            Some(rustls_pemfile::Item::RSAKey(key)) => return rustls::PrivateKey(key),
-            Some(rustls_pemfile::Item::PKCS8Key(key)) => return rustls::PrivateKey(key),
-            Some(rustls_pemfile::Item::ECKey(key)) => return rustls::PrivateKey(key),
+    loop
+    {
+        match rustls_pemfile::read_one(&mut reader).expect("cannot parse private key .pem file")
+        {
+            Some(rustls_pemfile::Item::RSAKey(key)) => return Ok(rustls::PrivateKey(key)),
+            Some(rustls_pemfile::Item::PKCS8Key(key)) => return Ok(rustls::PrivateKey(key)),
+            Some(rustls_pemfile::Item::ECKey(key)) => return Ok(rustls::PrivateKey(key)),
             None => break,
             _ => {}
         }
     }
 
-    panic!(
-        "no keys found in {:?} (encrypted keys not supported)",
-        filename
-    );
+    return Err(format!("no keys found in {:?} (encrypted keys not supported)", filename).into());
 }
 
-fn create_tls_config() -> Result<Arc<rustls::ServerConfig>, Box<dyn std::error::Error>>
+fn create_server_tls_config() -> Result<Arc<rustls::ServerConfig>, Box<dyn std::error::Error>>
 {
-    let roots = load_certs("ca.cert".into())?;
+    let roots = load_certs("cert/ec-cacert.pem".into())?;
     let mut client_auth_roots = RootCertStore::empty();
     for root in roots
     {
@@ -55,29 +55,58 @@ fn create_tls_config() -> Result<Arc<rustls::ServerConfig>, Box<dyn std::error::
 
     let versions : Vec<&'static rustls::SupportedProtocolVersion> = vec![&rustls::version::TLS13];
 
-    let certs = load_certs("server.cert")?;
+    let certs = load_certs("server.pem")?;
 
-    let privkey = load_private_key("server.key");
+    let privkey = load_private_key("server.key")?;
 
     let ocsp : Vec<u8> = vec![];
 
     let mut config = rustls::ServerConfig::builder()
         .with_cipher_suites(&suites)
         .with_safe_default_kx_groups()
-        .with_protocol_versions(versions.as_slice())
-        .expect("inconsistent cipher-suites/versions specified")
+        .with_protocol_versions(versions.as_slice())?
         .with_client_cert_verifier(client_auth)
-        .with_single_cert_with_ocsp_and_sct(certs, privkey, ocsp, vec![])
-        .expect("bad certificates/private key");
+        .with_single_cert_with_ocsp_and_sct(certs, privkey, ocsp, vec![])?;
 
     config.key_log = Arc::new(rustls::KeyLogFile::new());
 
     Ok(Arc::new(config))
 }
 
+pub fn create_client_tls_config() -> Result<Arc<rustls::ClientConfig>, Box<dyn std::error::Error>>
+{
+    let mut root_store = RootCertStore::empty(); 
+
+    let certfile = std::fs::File::open(&"cert/ec-cacert.pem")?;
+    let mut reader = BufReader::new(certfile);
+    root_store.add_parsable_certificates(&rustls_pemfile::certs(&mut reader)?);
+
+    let suites = vec![ 
+						 rustls::cipher_suite::TLS13_AES_128_GCM_SHA256, 
+						 rustls::cipher_suite::TLS13_AES_256_GCM_SHA384, 
+					 ];
+
+    let versions : Vec<&'static rustls::SupportedProtocolVersion> = vec![&rustls::version::TLS13];
+
+    let config = rustls::ClientConfig::builder()
+                 .with_cipher_suites(&suites)
+                 .with_safe_default_kx_groups()
+                 .with_protocol_versions(&versions)?
+                 .with_root_certificates(root_store);
+
+    let certs = load_certs("client.crt")?;
+    let key = load_private_key("client.key")?;
+
+    let mut conf = config.with_single_cert(certs, key)?;
+
+    conf.key_log = Arc::new(rustls::KeyLogFile::new());
+
+    Ok(Arc::new(conf))
+}
+
 pub fn load_configuration() -> Result<LoadBalancer, Box<dyn std::error::Error>>
 {
-	let tls_conf = create_tls_config()?;
+	let tls_conf = create_server_tls_config()?;
 
     let mut lb = LoadBalancer::new(tls_conf)?;
 
@@ -88,23 +117,17 @@ pub fn load_configuration() -> Result<LoadBalancer, Box<dyn std::error::Error>>
 
     let mut sg0 = ServerGroup::new(0);
 
-    sg0.server_addrs.insert(0, "127.0.0.1:2500".into());
-    sg0.server_health.insert(0, HealthChecker::new(0, "127.0.0.1:2500".into()));
-    sg0.server_addrs.insert(1, "127.0.0.1:2501".into());
-    sg0.server_health.insert(1, HealthChecker::new(1, "127.0.0.1:2501".into()));
-    sg0.server_addrs.insert(2, "127.0.0.1:2502".into());
-    sg0.server_health.insert(2, HealthChecker::new(2, "127.0.0.1:2502".into()));
+    sg0.add_server(0, "127.0.0.1:2500".into());
+    sg0.add_server(1, "127.0.0.1:2501".into());
+    sg0.add_server(2, "127.0.0.1:2502".into());
 
     lb.server_groups.insert(0, sg0);
     
     let mut sg1 = ServerGroup::new(1);
 
-    sg1.server_addrs.insert(3, "127.0.0.1:2500".into());
-    sg1.server_health.insert(3, HealthChecker::new(3, "127.0.0.1:2503".into()));
-    sg1.server_addrs.insert(4, "127.0.0.1:2501".into());
-    sg1.server_health.insert(4, HealthChecker::new(4, "127.0.0.1:2504".into()));
-    sg1.server_addrs.insert(5, "127.0.0.1:2502".into());
-    sg1.server_health.insert(5, HealthChecker::new(5, "127.0.0.1:2505".into()));
+    sg1.add_server(3, "127.0.0.1:2500".into());
+    sg1.add_server(4, "127.0.0.1:2501".into());
+    sg1.add_server(5, "127.0.0.1:2502".into());
 
     lb.server_groups.insert(0, sg1);
